requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
dvc
dvc[gs]
torch
torchvision
kornia
hydra-core
matplotlib
